src/backend/app/requirements.txt
Flask==2.3.0
mysql-connector-python==8.0.31
flask-cors
pandas>=1.5.0
numpy>=1.21.0
openpyxl==3.0.10
